+++ packages/winnow/test/filter.js
@@ -117,6 +117,21 @@ test('With an esri style envelope', (t) => {
   run('trees', options, 29744, t)
 })
 
+test('With an esri style envelope and wkt string for web mercator', (t) => {
+  const options = {
+    geometry: {
+      xmin: -13155799.066536672,
+      ymin: 4047806.77771083,
+      xmax: -13143569.142011061,
+      ymax: 4050673.16627152,
+      spatialReference: {
+        wkt: 'PROJCS["WGS_1984_Web_Mercator_Auxiliary_Sphere",GEOGCS["GCS_WGS_1984",DATUM["D_WGS_1984",SPHEROID["WGS_1984",6378137.0,298.257223563]],PRIMEM["Greenwich",0.0],UNIT["Degree",0.0174532925199433]],PROJECTION["Mercator_Auxiliary_Sphere"],PARAMETER["False_Easting",0.0],PARAMETER["False_Northing",0.0],PARAMETER["Central_Meridian",-97.03124999997486],PARAMETER["Standard_Parallel_1",0.0],PARAMETER["Auxiliary_Sphere_Type",0.0],UNIT["Meter",1.0]]'
+      }
+    }
+  }
+  run('trees', options, 29744, t)
+})
+
 test('With an empty multipolygon', (t) => {
   const options = {
     geometry: {